refactor(src/semantic-tokens.ts): renamed `src/semantic-tokens-simple.ts`
- removed old `**/semantic-tokens.*` file sin workspace and replaced
them with `src/semantic-token-simple.ts` (their temporary name in
history)
